Restructured the directories in the application
