+++ src/components/app/index.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import './styles.css';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="app">
+      </div>
+    );
+  }
+}
+
+export default App;
